pyproject.toml
[tool.poetry]
name = "repetition-generation"
version = "0.1.0"
description = ""
authors = ["Toshiki Kawamoto"]

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^1.0.2"
sumeval = "^0.2.2"
allennlp = "^2.9.3"
urllib3 = "^1.26.9"
mecab-python3 = "^1.0.5"
transformers = "^4.18.0"
importlib_metadata = "^4.11.3"
cached_property = "^1.5.2"
fugashi = "^1.1.2"
ipadic = "^1.0.0"
torch = {url = "https://download.pytorch.org/whl/cu115/torch-1.11.0%2Bcu115-cp37-cp37m-linux_x86_64.whl"}
pytorch-lightning = "^1.6.1"
Janome = "^0.4.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
